pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "google-search-results>=2.4.2",
  "openai>=2.29.0",
  "python-dotenv>=1.2.2",
  "requests>=2.32.5",
  "tavily-python>=0.7.23",
]

[tool.setuptools]
packages = ["ch01_simple_agent","ch02_llm_agent"]  # 只声明1个顶级包（核心！）
package-dir = {""= "src"}  # 包路径映射

[tool.uv]
package = true  # 启用包模式（现在结构合法，无冲突）